Introduces ObservedCadence class
